Request story 1: As a customer, I want to book a service and track its status.

+++ src/main/java/org/os/carcareservice/controller/CustomerCarController.java
@@ -25,7 +25,7 @@ public class CustomerCarController {
     // }
 
     @PostMapping("/customers/{id}/cars")
-    public CustomerCar addCarByCustomerId(@PathVariable Integer id, @RequestBody CustomerCar car) {
+    public CustomerCar addCarByCustomerId(@PathVariable Long id, @RequestBody CustomerCar car) {
         return carService.addCarByCustomerId(id, car);
     }
